backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
gunicorn==23.0.0
pytest==8.3.4
pytest-flask==1.3.0
